Moved FilterExpressionCheck to security package
