+++ elide-async/src/main/java/com/yahoo/elide/async/models/security/AsyncQueryInlineChecks.java
@@ -13,7 +13,7 @@ import com.yahoo.elide.core.filter.FilterPredicate;
 import com.yahoo.elide.core.filter.Operator;
 import com.yahoo.elide.core.filter.expression.FilterExpression;
 import com.yahoo.elide.security.ChangeSpec;
-import com.yahoo.elide.security.FilterExpressionCheck;
+import com.yahoo.elide.security.checks.FilterExpressionCheck;
 import com.yahoo.elide.security.RequestScope;
 import com.yahoo.elide.security.User;
 import com.yahoo.elide.security.checks.OperationCheck;
