+++ elide-core/src/test/java/com/yahoo/elide/parsers/expression/CanPaginateVisitorTest.java
@@ -18,7 +18,7 @@ import com.yahoo.elide.core.RequestScope;
 import com.yahoo.elide.core.TestDictionary;
 import com.yahoo.elide.core.filter.expression.FilterExpression;
 import com.yahoo.elide.security.ChangeSpec;
-import com.yahoo.elide.security.FilterExpressionCheck;
+import com.yahoo.elide.security.checks.FilterExpressionCheck;
 import com.yahoo.elide.security.User;
 import com.yahoo.elide.security.checks.Check;
 import com.yahoo.elide.security.checks.OperationCheck;
