+++ elide-integration-tests/src/test/java/example/FilterExpressionCheckObj.java
@@ -10,7 +10,7 @@ import com.yahoo.elide.annotation.ReadPermission;
 import com.yahoo.elide.core.Path;
 import com.yahoo.elide.core.filter.FilterPredicate;
 import com.yahoo.elide.core.filter.Operator;
-import com.yahoo.elide.security.FilterExpressionCheck;
+import com.yahoo.elide.security.checks.FilterExpressionCheck;
 import com.yahoo.elide.security.RequestScope;
 
 import java.util.ArrayList;
